+++ probdiffeq/strategies/filters.py
@@ -40,7 +40,6 @@ class FiSolution(NamedTuple):
 
     rv: Any
 
-    # todo: make a similar field in MarkovSequence
     num_data_points: float
 
 
@@ -58,7 +57,7 @@ class Filter(_strategy.Strategy[_FiState, Any]):
     def solution_from_tcoeffs(
         self, taylor_coefficients, *, num_data_points
     ) -> FiSolution:
-        ssv = self.implementation.extrapolation.init_state_space_var(
+        ssv = self.extrapolation.init_state_space_var(
             taylor_coefficients=taylor_coefficients
         )
         return FiSolution(ssv, num_data_points=num_data_points)
@@ -66,8 +65,6 @@ class Filter(_strategy.Strategy[_FiState, Any]):
     def extract(self, posterior: _FiState, /) -> FiSolution:
         return FiSolution(posterior.corrected, posterior.num_data_points)
 
-    # todo: make interpolation result into a named-tuple.
-    #  it is too confusing what those three posteriors mean.
     def case_right_corner(
         self, *, s0: _FiState, s1: _FiState, t, t0, t1, output_scale
     ) -> _collections.InterpRes[_FiState]:  # s1.t == t
@@ -131,7 +128,7 @@ class Filter(_strategy.Strategy[_FiState, Any]):
         return state.corrected.extract_qoi()
 
     def begin_extrapolation(self, posterior: _FiState, /, *, dt) -> _FiState:
-        extrapolate = self.implementation.extrapolation.begin_extrapolation
+        extrapolate = self.extrapolation.begin_extrapolation
         extrapolated = extrapolate(posterior.corrected, dt=dt)
         return _FiState(
             extrapolated=extrapolated,
@@ -139,11 +136,10 @@ class Filter(_strategy.Strategy[_FiState, Any]):
             num_data_points=posterior.num_data_points,
         )
 
-    # todo: make "output_extra" positional only. Then rename this mess.
     def begin_correction(
         self, output_extra: _FiState, /, *, vector_field, t, p
     ) -> Tuple[jax.Array, float, Any]:
-        return self.implementation.correction.begin_correction(
+        return self.correction.begin_correction(
             output_extra.extrapolated, vector_field=vector_field, t=t, p=p
         )
 
@@ -155,9 +151,8 @@ class Filter(_strategy.Strategy[_FiState, Any]):
         output_scale,
         state_previous: _FiState,
     ) -> _FiState:
-        extra = self.implementation.extrapolation
+        extra = self.extrapolation
         extrapolate_fn = extra.complete_extrapolation_without_reversal
-        # todo: extrapolation needs a serious signature-variable-renaming...
         ssv = extrapolate_fn(
             output_extra.extrapolated,
             s0=state_previous.corrected,
@@ -173,7 +168,7 @@ class Filter(_strategy.Strategy[_FiState, Any]):
     def complete_correction(
         self, extrapolated: _FiState, /, *, cache_obs
     ) -> Tuple[Any, Tuple[_FiState, Any]]:
-        obs, corr = self.implementation.correction.complete_correction(
+        obs, corr = self.correction.complete_correction(
             extrapolated=extrapolated.extrapolated, cache=cache_obs
         )
         corr = _FiState(