+++ probdiffeq/test_util.py
@@ -16,35 +16,29 @@ def generate_solver(
 
     Examples
     --------
+    >>> import jax.numpy as jnp
     >>> from jax.config import config
     >>> config.update("jax_platform_name", "cpu")
+    >>> jnp.set_printoptions(suppress=True, precision=2)  # summarise arrays
 
     >>> from probdiffeq import ivpsolvers
     >>> from probdiffeq.statespace import recipes
     >>> from probdiffeq.strategies import smoothers
 
     >>> print(generate_solver())
-    MLESolver(strategy=Filter(implementation=<Implementation with num_derivatives=4, ode_order=1>))
+    MLESolver(Filter(<Isotropic IBM with num_derivatives=4>, <TS0 with ode_order=1>))
 
     >>> print(generate_solver(num_derivatives=1))
-    MLESolver(strategy=Filter(implementation=<Implementation with num_derivatives=1, ode_order=1>))
+    MLESolver(Filter(<Isotropic IBM with num_derivatives=1>, <TS0 with ode_order=1>))
 
     >>> print(generate_solver(solver_factory=ivpsolvers.DynamicSolver))
-    DynamicSolver(strategy=Filter(implementation=<Implementation with num_derivatives=4, ode_order=1>))
+    DynamicSolver(Filter(<Isotropic IBM with num_derivatives=4>, <TS0 with ode_order=1>))
 
     >>> impl_fcty = recipes.ts1_dense
     >>> strat_fcty = smoothers.Smoother
     >>> print(generate_solver(strategy_factory=strat_fcty, impl_factory=impl_fcty, ode_shape=(1,)))  # noqa: E501
-    MLESolver(strategy=Smoother(implementation=<Implementation with num_derivatives=4, ode_order=1>))
+    MLESolver(Smoother(<Dense IBM with num_derivatives=4, ode_shape=(1,)>, <TS1 with ode_order=1>))
     """
-    impl = impl_factory(**impl_factory_kwargs)
-    strat = strategy_factory(impl)
-
+    extra, correct = impl_factory(**impl_factory_kwargs)
+    strat = strategy_factory(extra, correct)
     return solver_factory(strat)
-    # # I am not too happy with the need for this distinction below...
-    #
-    # if solver_factory in [ivpsolvers.MLESolver, ivpsolvers.DynamicSolver]:
-    #     return solver_factory(strat)
-    #
-    # output_scale = impl.extrapolation.init_output_scale()
-    # return solver_factory(strat, output_scale=output_scale)